requirements.txt
Scrapy==2.8.0
pytz==2021.1
psycopg2==2.9.5
python-dotenv==1.0.0